Outsourced GraphSearch from HomotopyClassPlanner to own class.
